+++ middleware/addressRequired.js
@@ -1,5 +1,5 @@
 export default function ({ store, redirect }) {
   if (!store.state.session) {
-    return redirect('/address')
+    return redirect('/welcome')
   }
 }
