+++ middleware/localSigning.js
@@ -2,6 +2,6 @@ import network from '~/common/network'
 
 export default function ({ store, redirect }) {
   if (!network.localSigning) {
-    return redirect('/address')
+    return redirect('/welcome')
   }
 }
