--- pages/address/index.vue
@@ -1,9 +1,0 @@
-<template>
-  <CardSignInRequired />
-</template>
-
-<script>
-export default {
-  name: `page-address`,
-}
-</script>
