+++ pages/signin/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <SessionFrame :icon="`lock`">
+  <div>
     <Form :submit="onSubmit" class="session-container">
       <h2 class="session-title">Sign in with account</h2>
       <div class="session-main bottom-indent">
@@ -47,7 +47,7 @@
         <Button value="Sign In" :disabled="loading" />
       </div>
     </Form>
-  </SessionFrame>
+  </div>
 </template>
 
 <script>
@@ -56,6 +56,7 @@ import { getWallet, getWalletIndex } from '~/common/keystore'
 
 export default {
   name: `sign-in`,
+  layout: 'session',
   data: () => ({
     signInAddress: undefined,
     signInPassword: ``,
